requirements.txt
dj-database-url==0.4.2
Django==1.10.3
django-filter==1.1.0
django-multiselectfield==0.1.8
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
